Create NormCacheTest.cpp
Tests the ability of a EuclideanVector container which caches the norm to consistently generate the correct norm, despite having it's magnitude values changed in various ways (I.e. the flag which tells the EV class it needs to re-calculate the norm are never forgotten in setter methods).

The last step tests the speed difference between using the cached value, and constantly having to re-calculate it. Although this test does not tell us a whole lot, it is interesting to see the difference.